Add code suggestions about move data from activity to view model AND-12732 Text editing

+++ app/src/main/java/mega/privacy/android/app/textFileEditor/TextFileEditorViewModel.kt
@@ -5,9 +5,9 @@ import android.app.ActivityManager
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.os.Bundle
 import androidx.core.net.toUri
 import androidx.hilt.lifecycle.ViewModelInject
+import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
@@ -15,7 +15,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import mega.privacy.android.app.AndroidCompletedTransfer
-import mega.privacy.android.app.MegaApplication
+import mega.privacy.android.app.DatabaseHandler
 import mega.privacy.android.app.UploadService
 import mega.privacy.android.app.arch.BaseRxViewModel
 import mega.privacy.android.app.di.MegaApi
@@ -27,6 +27,7 @@ import mega.privacy.android.app.utils.ChatUtil.authorizeNodeIfPreview
 import mega.privacy.android.app.utils.Constants.*
 import mega.privacy.android.app.utils.FileUtil.getLocalFile
 import mega.privacy.android.app.utils.FileUtil.isFileAvailable
+import mega.privacy.android.app.utils.LogUtil
 import mega.privacy.android.app.utils.LogUtil.logError
 import mega.privacy.android.app.utils.MegaNodeUtil.handleSelectFolderToCopyResult
 import mega.privacy.android.app.utils.MegaNodeUtil.handleSelectFolderToImportResult
@@ -42,7 +43,8 @@ import java.net.URL
 class TextFileEditorViewModel @ViewModelInject constructor(
     @MegaApi private val megaApi: MegaApiAndroid,
     @MegaApiFolder private val megaApiFolder: MegaApiAndroid,
-    private val megaChatApi: MegaChatApiAndroid
+    private val megaChatApi: MegaChatApiAndroid,
+    private val dbH: DatabaseHandler
 ) : BaseRxViewModel() {
 
     companion object {
@@ -50,92 +52,104 @@ class TextFileEditorViewModel @ViewModelInject constructor(
         const val CREATE_MODE = "CREATE_MODE"
         const val VIEW_MODE = "VIEW_MODE"
         const val EDIT_MODE = "EDIT_MODE"
-
-        const val CONTENT_TEXT = "CONTENT_TEXT"
     }
 
-    private lateinit var api: MegaApiAndroid
+    private val textFileEditorData: MutableLiveData<TextFileEditorData> =
+        MutableLiveData(TextFileEditorData())
+    private val mode: MutableLiveData<String> = MutableLiveData()
+    private val savingMode: MutableLiveData<String> = MutableLiveData()
+    private val fileName: MutableLiveData<String> = MutableLiveData()
+    private val contentText: MutableLiveData<String> = MutableLiveData()
+    private val editedText: MutableLiveData<String> by lazy { MutableLiveData<String>() }
+    private val editableAdapter: MutableLiveData<Boolean> = MutableLiveData()
+    private val creationOrEditionSuccess: MutableLiveData<Boolean> = MutableLiveData()
 
-    private var fileName = ""
-    private var node: MegaNode? = null
-    private var fileUri: Uri? = null
-    private var fileSize: Long? = null
-    private var mode = VIEW_MODE
-    private var adapterType: Int = INVALID_VALUE
-    private var msgChat: MegaChatMessage? = null
-    private var chatRoom: MegaChatRoom? = null
+    fun onSavingMode(): LiveData<String> = savingMode
 
-    private var needStopHttpServer = false
+    fun getFileName(): LiveData<String> = fileName
 
-    private val contentText: MutableLiveData<String> by lazy { MutableLiveData<String>() }
     fun onContentTextRead(): LiveData<String> = contentText
 
-    fun getContentText(): String? = contentText.value
+    fun isEditableAdapter(): LiveData<Boolean> = editableAdapter
 
+    fun onCreationOrEditionFinished(): LiveData<Boolean> = creationOrEditionSuccess
 
-    fun setContentText(editedContentText: String) {
-        contentText.value = editedContentText
-    }
+    fun getNode(): MegaNode? = textFileEditorData.value?.node
 
-    fun getFileName(): String = fileName
-
-    fun getNode(): MegaNode? = node
+    fun getNodeAccess(): Int = megaApi.getAccess(getNode())
 
     fun updateNode() {
-        if (node == null) return
+        val node = textFileEditorData.value?.node ?: return
 
-        node = megaApi.getNodeByHandle(node!!.handle)
+        textFileEditorData.value?.node = megaApi.getNodeByHandle(node.handle)
     }
 
-    fun updateNode(handle: Long) {
-        node = megaApi.getNodeByHandle(handle)
-    }
+    fun getFileUri(): Uri? = textFileEditorData.value?.fileUri
 
-    fun getFileUri(): Uri = fileUri!!
+    fun getFileSize(): Long? = textFileEditorData.value?.fileSize
 
-    fun getFileSize(): Long = fileSize!!
+    fun getAdapterType(): Int = textFileEditorData.value?.adapterType ?: INVALID_VALUE
 
-    fun getNodeAccess(): Int = megaApi.getAccess(node)
+    fun getMsgChat(): MegaChatMessage? = textFileEditorData.value?.msgChat
 
-    fun getMode(): String = mode
+    fun getChatRoom(): MegaChatRoom? = textFileEditorData.value?.chatRoom
 
-    fun isViewMode(): Boolean = mode == VIEW_MODE
+    fun getMode(): LiveData<String> = mode
 
-    fun setViewMode() {
-        mode = VIEW_MODE
+    fun isViewMode(): Boolean = mode.value == VIEW_MODE
+
+    private fun setViewMode() {
+        mode.value = VIEW_MODE
     }
 
     fun setEditMode() {
-        mode = EDIT_MODE
+        mode.value = EDIT_MODE
     }
 
-    fun getAdapterType(): Int = adapterType
+    fun isSavingMode(): Boolean = savingMode.value != null
 
-    fun getMsgChat(): MegaChatMessage? = msgChat
+    fun isSavingModeEdit(): Boolean = savingMode.value == EDIT_MODE
 
-    fun getChatRoom(): MegaChatRoom? = chatRoom
+    fun resetSavingMode() {
+        savingMode.value = null
+    }
+
+    fun getEditedText(): String? = editedText.value
+
+    fun setEditedText(text: String?) {
+        editedText.value = text
+    }
+
+    fun getNameOfFile(): String = fileName.value ?: ""
+
+    fun removeCreationOrEditionSuccessObservers(owner: LifecycleOwner) {
+        creationOrEditionSuccess.removeObservers(owner)
+        creationOrEditionSuccess.value = false
+    }
 
     /**
      * Checks if the file can be editable depending on the current adapter.
-     *
-     * @return True if the file can be editable, false otherwise.
      */
-    fun isEditableAdapter(): Boolean = adapterType != OFFLINE_ADAPTER
-            && adapterType != RUBBISH_BIN_ADAPTER && !megaApi.isInRubbish(node)
-            && adapterType != FILE_LINK_ADAPTER
-            && adapterType != FOLDER_LINK_ADAPTER
-            && adapterType != ZIP_ADAPTER
-            && adapterType != FROM_CHAT
-            && (getNodeAccess() == MegaShare.ACCESS_OWNER || getNodeAccess() == MegaShare.ACCESS_READWRITE)
+    private fun setEditableAdapter() {
+        editableAdapter.value = getAdapterType() != OFFLINE_ADAPTER
+                && getAdapterType() != RUBBISH_BIN_ADAPTER && !megaApi.isInRubbish(getNode())
+                && getAdapterType() != FILE_LINK_ADAPTER
+                && getAdapterType() != FOLDER_LINK_ADAPTER
+                && getAdapterType() != ZIP_ADAPTER
+                && getAdapterType() != FROM_CHAT
+                && getAdapterType() != INVALID_VALUE
+                && (getNodeAccess() == MegaShare.ACCESS_OWNER || getNodeAccess() == MegaShare.ACCESS_READWRITE)
+    }
 
     /**
-     * Gets all necessary values from intent and savedInstanceState if available.
+     * Gets all necessary values from intent if available.
      *
-     * @param intent             Received intent.
-     * @param savedInstanceState Saved state.
+     * @param intent Received intent.
      */
-    fun setValuesFromIntent(intent: Intent, savedInstanceState: Bundle?) {
-        adapterType = intent.getIntExtra(INTENT_EXTRA_KEY_ADAPTER_TYPE, INVALID_VALUE)
+    fun setValuesFromIntent(intent: Intent) {
+        val adapterType = intent.getIntExtra(INTENT_EXTRA_KEY_ADAPTER_TYPE, INVALID_VALUE)
+        textFileEditorData.value?.adapterType = adapterType
+        setEditableAdapter()
 
         when (adapterType) {
             FROM_CHAT -> {
@@ -143,16 +157,18 @@ class TextFileEditorViewModel @ViewModelInject constructor(
                 val chatId = intent.getLongExtra(CHAT_ID, MEGACHAT_INVALID_HANDLE)
 
                 if (msgId != MEGACHAT_INVALID_HANDLE && chatId != MEGACHAT_INVALID_HANDLE) {
-                    chatRoom = megaChatApi.getChatRoom(chatId)
-                    msgChat = megaChatApi.getMessage(chatId, msgId)
+                    textFileEditorData.value?.chatRoom = megaChatApi.getChatRoom(chatId)
+                    var msgChat = megaChatApi.getMessage(chatId, msgId)
 
                     if (msgChat == null) {
                         msgChat = megaChatApi.getMessageFromNodeHistory(chatId, msgId)
                     }
 
                     if (msgChat != null) {
-                        node = authorizeNodeIfPreview(
-                            msgChat!!.megaNodeList.get(0),
+                        textFileEditorData.value?.msgChat = msgChat
+
+                        textFileEditorData.value?.node = authorizeNodeIfPreview(
+                            msgChat.megaNodeList.get(0),
                             megaChatApi,
                             megaApi,
                             chatId
@@ -162,24 +178,28 @@ class TextFileEditorViewModel @ViewModelInject constructor(
             }
             OFFLINE_ADAPTER, ZIP_ADAPTER -> {
                 val filePath = intent.getStringExtra(INTENT_EXTRA_KEY_PATH)
-                fileUri = filePath!!.toUri()
-                fileSize = File(filePath).length()
+
+                if (filePath != null) {
+                    textFileEditorData.value?.fileUri = filePath.toUri()
+                    textFileEditorData.value?.fileSize = File(filePath).length()
+                }
             }
             FILE_LINK_ADAPTER -> {
-                node = MegaNode.unserialize(intent.getStringExtra(EXTRA_SERIALIZE_STRING))
+                textFileEditorData.value?.node =
+                    MegaNode.unserialize(intent.getStringExtra(EXTRA_SERIALIZE_STRING))
             }
             FOLDER_LINK_ADAPTER -> {
-                node = megaApiFolder.getNodeByHandle(
+                val node = megaApiFolder.getNodeByHandle(
                     intent.getLongExtra(
                         INTENT_EXTRA_KEY_HANDLE,
                         INVALID_HANDLE
                     )
                 )
 
-                node = megaApiFolder.authorizeNode(node)
+                textFileEditorData.value?.node = megaApiFolder.authorizeNode(node)
             }
             else -> {
-                node = megaApi.getNodeByHandle(
+                textFileEditorData.value?.node = megaApi.getNodeByHandle(
                     intent.getLongExtra(
                         INTENT_EXTRA_KEY_HANDLE,
                         INVALID_HANDLE
@@ -188,16 +208,12 @@ class TextFileEditorViewModel @ViewModelInject constructor(
             }
         }
 
-        api = if (adapterType == FOLDER_LINK_ADAPTER) megaApiFolder else megaApi
+        textFileEditorData.value?.api =
+            if (adapterType == FOLDER_LINK_ADAPTER) megaApiFolder else megaApi
 
-        if (savedInstanceState != null) {
-            mode = savedInstanceState.getString(MODE) ?: VIEW_MODE
-            contentText.value = savedInstanceState.getString(CONTENT_TEXT)
-        } else {
-            mode = intent.getStringExtra(MODE) ?: VIEW_MODE
-        }
+        mode.value = intent.getStringExtra(MODE) ?: VIEW_MODE
 
-        fileName = intent.getStringExtra(INTENT_EXTRA_KEY_FILE_NAME) ?: node?.name!!
+        fileName.value = intent.getStringExtra(INTENT_EXTRA_KEY_FILE_NAME) ?: getNode()?.name!!
     }
 
     /**
@@ -218,8 +234,8 @@ class TextFileEditorViewModel @ViewModelInject constructor(
     private suspend fun readFile(mi: ActivityManager.MemoryInfo) {
         withContext(Dispatchers.IO) {
             val localFileUri =
-                if (adapterType == OFFLINE_ADAPTER || adapterType == ZIP_ADAPTER) fileUri.toString()
-                else getLocalFile(null, node?.name, node?.size!!)
+                if (getAdapterType() == OFFLINE_ADAPTER || getAdapterType() == ZIP_ADAPTER) getFileUri().toString()
+                else getLocalFile(null, getNode()?.name, getNode()?.size!!)
 
             if (!isTextEmpty(localFileUri)) {
                 val localFile = File(localFileUri)
@@ -230,9 +246,11 @@ class TextFileEditorViewModel @ViewModelInject constructor(
                 }
             }
 
+            val api = textFileEditorData.value?.api ?: return@withContext
+
             if (api.httpServerIsRunning() == 0) {
                 api.httpServerStart()
-                needStopHttpServer = true
+                textFileEditorData.value?.needStopHttpServer = true
             }
 
             api.httpServerSetMaxBufferSize(
@@ -240,10 +258,10 @@ class TextFileEditorViewModel @ViewModelInject constructor(
                 else MAX_BUFFER_16MB
             )
 
-            val uri = api.httpServerGetLocalLink(node)
+            val uri = api.httpServerGetLocalLink(getNode())
             if (uri == null) {
                 logError("Error getting the file uri.")
-                contentText.value = ""
+                return@withContext
             }
 
             val url = URL(uri)
@@ -275,6 +293,7 @@ class TextFileEditorViewModel @ViewModelInject constructor(
             }
 
             contentText.postValue(sb.toString())
+            editedText.postValue(sb.toString())
         }
     }
 
@@ -282,71 +301,59 @@ class TextFileEditorViewModel @ViewModelInject constructor(
      * Starts the save file content action by creating a temp file, setting the new or modified text,
      * and then uploading it to the Cloud.
      *
-     * @param context     Current context.
-     * @param contentText The new or modified content text.
+     * @param context Current context.
      */
-    fun saveFile(context: Context, contentText: String): Boolean {
-        val tempFile = buildTempFile(context, fileName)
+    fun saveFile(context: Context) {
+        if (!isFileEdited()) {
+            setViewMode()
+            return
+        }
+
+        val tempFile = buildTempFile(context, fileName.value)
         if (tempFile == null) {
             logError("Cannot get temporal file.")
-            return false
+
+            return
         }
 
         val fileWriter = FileWriter(tempFile.absolutePath)
         val out = BufferedWriter(fileWriter)
-        out.write(contentText)
+        out.write(editedText.value)
         out.close()
 
         if (!isFileAvailable(tempFile)) {
             logError("Cannot manage temporal file.")
-            return false
+            return
         }
 
         val uploadIntent = Intent(context, UploadService::class.java)
             .putExtra(UploadService.EXTRA_UPLOAD_TXT, true)
             .putExtra(UploadService.EXTRA_FILEPATH, tempFile.absolutePath)
-            .putExtra(UploadService.EXTRA_NAME, fileName)
+            .putExtra(UploadService.EXTRA_NAME, fileName.value)
             .putExtra(UploadService.EXTRA_SIZE, tempFile.length())
             .putExtra(
                 UploadService.EXTRA_PARENT_HASH,
-                if (mode == CREATE_MODE && node == null) {
+                if (mode.value == CREATE_MODE && getNode() == null) {
                     megaApi.rootNode.handle
-                } else if (mode == CREATE_MODE && node != null) {
-                    node?.handle
+                } else if (mode.value == CREATE_MODE) {
+                    getNode()?.handle
                 } else {
-                    node?.parentHandle
+                    getNode()?.parentHandle
                 }
             )
 
         context.startService(uploadIntent)
-
-        return true
+        savingMode.value = mode.value
+        mode.value = VIEW_MODE
     }
 
     /**
-     * Stops the http server if it has been started before.
+     * Stops the http server if has been started before.
      */
     fun checkIfNeedsStopHttpServer() {
-        if (needStopHttpServer) {
-            api.httpServerStop()
+        if (textFileEditorData.value?.needStopHttpServer == true) {
+            textFileEditorData.value?.api?.httpServerStop()
         }
-    }
-
-    /**
-     * Checks if the completed transfer refers to the same node of current view.
-     *
-     * @param completedTransfer Completed transfer to check.
-     * @return True if the completed transfer refers to the same node, false otherwise.
-     */
-    fun isSameNode(completedTransfer: AndroidCompletedTransfer): Boolean {
-        val fileParentHandle = when {
-            node == null -> megaApi.rootNode.handle
-            node!!.isFolder -> node!!.handle
-            else -> node!!.parentHandle
-        }
-
-        return completedTransfer.fileName == fileName
-                && completedTransfer.parentHandle == fileParentHandle
     }
 
     /**
@@ -379,7 +386,7 @@ class TextFileEditorViewModel @ViewModelInject constructor(
                 handleSelectFolderToImportResult(
                     resultCode,
                     toHandle,
-                    node!!,
+                    getNode()!!,
                     snackbarShower,
                     activityLauncher
                 )
@@ -398,5 +405,63 @@ class TextFileEditorViewModel @ViewModelInject constructor(
                 activityLauncher
             )
         }
+    }
+
+    /**
+     * Checks if the content of the file has been modified.
+     *
+     * @return True if the content has been modified, false otherwise.
+     */
+    fun isFileEdited(): Boolean = contentText.value != editedText.value
+
+    /**
+     * Checks if the completed transfer refers to the same node of current view.
+     *
+     * @param completedTransfer Completed transfer to check.
+     * @return True if the completed transfer refers to the same getNode(), false otherwise.
+     */
+    private fun isSameNode(completedTransfer: AndroidCompletedTransfer): Boolean {
+        val fileParentHandle = when {
+            getNode() == null -> megaApi.rootNode.handle
+            getNode()!!.isFolder -> getNode()!!.handle
+            else -> getNode()!!.parentHandle
+        }
+
+        return completedTransfer.fileName == fileName.value
+                && completedTransfer.parentHandle == fileParentHandle
+    }
+
+    /**
+     * Finishes the create or edit action if the received completed transfer makes reference to
+     * the current opened file.
+     *
+     * @param completedTransferId The completed transfer identifier.
+     */
+    fun finishCreationOrEdition(completedTransferId: Long) {
+        if (completedTransferId == INVALID_ID.toLong()) {
+            LogUtil.logWarning("Invalid completedTransferId")
+            return
+        }
+
+        val completedTransfer = dbH.getcompletedTransfer(completedTransferId)
+        if (completedTransfer == null) {
+            LogUtil.logWarning("Invalid completedTransfer")
+            return
+        }
+
+        if (!isSameNode(completedTransfer)) {
+            LogUtil.logWarning("Not the same file, no update needed.")
+            return
+        }
+
+        if (completedTransfer.state != MegaTransfer.STATE_COMPLETED) {
+            creationOrEditionSuccess.value = false
+            return
+        }
+
+        contentText.value = editedText.value
+        textFileEditorData.value?.node =
+            megaApi.getNodeByHandle(completedTransfer.nodeHandle.toLong())
+        creationOrEditionSuccess.value = true
     }
 }
